backend/requirements.txt
django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
mysqlclient==2.0.3
setuptools
pillow>=11.3.0
